# docs/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Heimdall federated analysis API",
    "description": "Heimdall HTTP API -- the contract the researcher UI (frontend/) talks to.\n\n    SERVER_OUT=server/api_out uvicorn server.api:app --port 8500 --workers 1\n\nStart mock TREs separately with `python scripts/dev_tres.py`. Neither TREs nor\nFLARE are required for startup or /health. API_TRE_TIMEOUT (seconds per HTTP\noperation, default 10) and API_TRE_WORKERS (default 8) control adapter fan-out;\nAPI_CORS_ORIGINS (comma-separated, default http://localhost:3000) allows the UI\norigin. Use an exclusive SERVER_OUT directory; see analysis_service's\nconcurrency and demo-versus-secure-deployment notes. Variable names are\ncanonical (snp_rs001), never local column names; requests carry variable\nnames, never participant data. Response shapes: server/schemas.py, exported to\ndocs/schemas/ and served at /openapi.json.\n",
    "version": "0.2.0"
  },
  "paths": {
...
    "/allele-frequency": {
      "get": {
        "summary": "Allele Frequency",
        "operationId": "allele_frequency_allele_frequency_get",
        "parameters": [
          {
            "name": "variant",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Variant"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Project Id"
            }
          },
          {
            "name": "min_cell_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 5,
              "default": 5,
              "title": "Min Cell Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/linear-regression": {
      "post": {
        "summary": "Linear Regression",
        "operationId": "linear_regression_linear_regression_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegressionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/logistic-regression": {
      "post": {
        "summary": "Logistic Regression",
        "description": "Federated logistic regression (Newton-Raphson/IRLS). Unlike /linear-regression this\nmakes up to `rounds` sequential round-trips to every site, so it is not one call.",
        "operationId": "logistic_regression_logistic_regression_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogisticRegressionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/metadata": {
      "get": {
        "summary": "Metadata",
        "operationId": "metadata_metadata_get",
        "parameters": [
          {
            "name": "probe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "probe each TRE's /health (2 s timeout)",
              "default": true,
              "title": "Probe"
            },
            "description": "probe each TRE's /health (2 s timeout)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Metadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/examples/{name}": {
      "get": {
        "summary": "Example",
        "operationId": "example_examples__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/run": {
      "post": {
        "summary": "Submit Run",
        "operationId": "submit_run_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/run/{run_id}": {
      "get": {
        "summary": "Run Status",
        "operationId": "run_status_run__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/overseer/{spec_hash}/{decision}": {
      "post": {
        "summary": "Overseer Decide",
        "operationId": "overseer_decide_overseer__spec_hash___decision__post",
        "parameters": [
          {
            "name": "spec_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spec Hash"
            }
          },
          {
            "name": "decision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Decision"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/OverseerDecision"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audit": {
      "get": {
        "summary": "Audit",
        "description": "Per-site audit lines (newest first). AUDIT_ROOT/<tre_id>/audit.jsonl, default <repo>/audit/;\nAUDIT_DIR (single directory) is honoured for the in-process demo layout.",
        "operationId": "audit_audit_get",
        "parameters": [
          {
            "name": "spec_hash",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Spec Hash"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalysisResponse": {
        "properties": {
          "spec_hash": {
            "type": "string",
            "title": "Spec Hash"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "flagged",
              "failed"
            ],
            "title": "Status"
          },
          "decision": {
            "type": "string",
            "enum": [
              "OK",
              "FLAGGED"
            ],
            "title": "Decision"
          },
          "reasons": {
            "items": {
              "type": "string",
              "enum": [
                "min_sites",
                "site_suppression",
                "k_anon",
                "dominance",
                "differencing",
                "analysis",
                "disclosure_review_required"
              ]
            },
            "type": "array",
            "title": "Reasons"
          },
          "sites_expected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sites Expected"
          },
          "sites_reported": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sites Reported"
          },
          "sites_missing": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sites Missing"
          },
          "sites_failed": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Sites Failed"
          },
          "coverage": {
            "type": "string",
            "title": "Coverage"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReleasedResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spec_hash",
          "status",
          "decision",
          "reasons",
          "sites_expected",
          "sites_reported",
          "sites_missing",
          "sites_failed",
          "coverage"
        ],
        "title": "AnalysisResponse",
        "description": "Response of the analysis endpoints. `result` is present only when the\ndisclosure check passed; flagged runs wait in the overseer queue."
      },
...
      "AuditLog": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AuditLog"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogisticRegressionRequest": {
        "properties": {
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Features"
          },
          "target": {
            "type": "string",
            "minLength": 1,
            "title": "Target"
          },
          "project_id": {
            "type": "string",
            "minLength": 1,
            "title": "Project Id"
          },
          "filters": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Filters"
          },
          "min_cell_size": {
            "type": "integer",
            "minimum": 5.0,
            "title": "Min Cell Size",
            "default": 5
          },
          "rounds": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Rounds",
            "default": 25
          },
          "tol": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Tol",
            "default": 1e-08
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "features",
          "target",
          "project_id"
        ],
        "title": "LogisticRegressionRequest",
        "description": "POST /logistic-regression: a RegressionRequest plus the Newton-Raphson budget."
      },
...
      "Metadata": {
        "properties": {
          "variables": {
            "items": {
              "$ref": "#/components/schemas/VariableInfo"
            },
            "type": "array",
            "title": "Variables"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectInfo"
            },
            "type": "array",
            "title": "Projects"
          },
          "examples": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Examples"
          },
          "sites": {
            "items": {
              "$ref": "#/components/schemas/SiteInfo"
            },
            "type": "array",
            "title": "Sites"
          },
          "analysis_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Analysis Types"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "variables",
          "projects",
          "examples",
          "sites",
          "analysis_types"
        ],
        "title": "Metadata"
      },
...
      "OverseerDecision": {
        "properties": {
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          },
          "by": {
            "type": "string",
            "title": "By",
            "default": "overseer-ui"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OverseerDecision"
      },
...
      "RegressionRequest": {
        "properties": {
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Features"
          },
          "target": {
            "type": "string",
            "minLength": 1,
            "title": "Target"
          },
          "project_id": {
            "type": "string",
            "minLength": 1,
            "title": "Project Id"
          },
          "filters": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Filters"
          },
          "min_cell_size": {
            "type": "integer",
            "minimum": 5.0,
            "title": "Min Cell Size",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "features",
          "target",
          "project_id"
        ],
        "title": "RegressionRequest",
        "description": "Convenience body for POST /linear-regression (same semantics as an AnalysisSpec)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}